Cave Music and secret Music Player update 1.11
Added Music to the Cave of Profound Misery and a secret Music Player.  Quite happy to work out a method using two variables (cave and trigger) so that you don't reset the music every time you step on the spaces and only when you actually enter or leave the cave.  EG:  Only if trigger ==1 does it reset back ground music but it only becomes trigger =1 when you enter the cave and the cave music is invoked.  Similar mechanics so that the only time bg music is reset if cave music is already playing.